feat(Icon): move the interface of Icon to index
